+++ pyqed/mps/__init__.py
@@ -1,2 +1,3 @@
 from .mps import *
+from .dmrg import DMRG
 # from .fermion import *
